last changes

+++ esbuild.config.mjs
@@ -8,10 +8,12 @@ const renamePlugin = () => ({
 	name: 'rename-plugin',
 	setup(build) {
 		build.onEnd(async () => {
-			try {
-				fs.renameSync('./main.css', './styles.css')
-			} catch (e) {
-				console.error('Failed to rename file:', e)
+			if (fs.existsSync("./main.css")) {
+				try {
+					fs.renameSync('./main.css', './styles.css')
+				} catch (e) {
+					console.error('Failed to rename file:', e)
+				}
 			}
 		});
 	},
@@ -20,15 +22,10 @@ const renamePlugin = () => ({
 const prod = (process.argv[2] === "production")
 
 const context = await esbuild.context({
-	// banner: {
-	// 	js: banner,
-	// },
 	entryPoints: ["main.tsx"],
 	bundle: true,
 	plugins: [
-		sassPlugin({
-			type: "css"
-		}),
+		sassPlugin(),
 		renamePlugin(),
 	],
 	external: [
